feat: abstract out a VoiceProvider
The original TTSService is still here in the OldTTSService for reference.
VoiceProviders abstract away the underlying Voice generation not only based on OS but also on tools. E.g. It would be easy to add a PiperVoiceProvider if desired.
Also added a Concept of AudioSinks which abstract away how to play sound based on a given OS.
A factory is provided to construct a given TTSService based on the current system.
All things considered this is pretty crude for the moment but is more meant as a Draft on how things could look. Also the new TTSSerivce contains a bunch of old functions that are just stubs to make the thing compile.
The test given is also more of a smoke test to see if this works.

Let me know what you think